Extract hook useFilterEmptyQueryParams to a file
